# neon: add dotprod dispatch implementation