… cmake/OpenCVCompilerOptimizations.cmake
… modules/core/CMakeLists.txt