* read vector at runtime
 * add enum

dot product instruction is an extended instruction of NEON on Aarch64.
I've used dispatch framework so it won't harm the other platforms.
Performance are reasonable

Topics I'd like to ask specifically for review
- The way to extract dot product instruction
  -  i.e.: it could be possible that other architecture hires dot product.
  - Is it OK to define `DOTPROD` for Armv8?
- The name, especially the `enum`, and the `DOTPROD` which gets shown for the config
  - I reused the keyword `CV_NEON_DOT` but will be there a better way?
- Support in Clang and MSVC (i.e.: I only checked on GCC)

Here's the performance summary
Jetson Orin and RK3588 support this dot product instructions.
Other platforms don't, but it's not making a harm.

|Board | Name of Test|before|after|after vs before (x-factor)|
|---|---|:-:|:-:|:-:|
|Jetson Orin|dot::MatType_Length::(8UC1, 1024)|0.235 ms|0.073 ms|3.20|
|Jetson Orin|dot::MatType_Length::(8SC1, 1024)|0.170 ms|0.073 ms|2.33|
| RK3588    |dot::MatType_Length::(8UC1, 1024)|0.269 ms|0.084 ms|3.22|
| RK3588    |dot::MatType_Length::(8SC1, 1024)|0.196 ms|0.069 ms|2.85|
|Raspberry Pi 4|dot::MatType_Length::(8UC1, 1024)|0.581 ms|0.583 ms|1.00|
|Raspberry Pi 4|dot::MatType_Length::(8SC1, 1024)|0.571 ms|0.569 ms|1.00|
|Jetson Nano|dot::MatType_Length::(8UC1, 1024)|0.743 ms|0.738 ms|1.01|
|Jetson Nano|dot::MatType_Length::(8SC1, 1024)|0.675 ms|0.666 ms|1.01|
|Jetson Xavier|dot::MatType_Length::(8UC1, 1024)|0.492 ms|0.528 ms|0.93|
|Jetson Xavier|dot::MatType_Length::(8SC1, 1024)|0.389 ms|0.423 ms|0.92|


I also confirmed that this PR doesn't harm the Arm 32bit build.

### Pull Request Readiness Checklist

See details at https://github.com/opencv/opencv/wiki/How_to_contribute#making-a-good-pull-request

- [x] I agree to contribute to the project under Apache 2 License.
- [x] To the best of my knowledge, the proposed patch is not based on a code under GPL or another license that is incompatible with OpenCV
- [x] The PR is proposed to the proper branch
- [ ] There is a reference to the original bug report and related work
- [x] There is accuracy test, performance test and test data in opencv_extra repository, if applicable
      Patch to opencv_extra has the same branch name.
- [ ] The feature is well documented and sample code can be built with the project CMake